chore: Move some code
required_action_url and get_next_url are better in utils/urls.py
